Add support for executing commands in the agent container
Introduced a new option `-b <branch_name>` to specify the branch name.
Updated the usage message to reflect the new option and added support for
executing commands in the container. The script now parses command-line
options to determine the branch name and any commands to execute. If a
command is provided, it will be executed in the container, and the
container will exit afterward. Updated the Docker run command to pass any
provided commands to the container.

Signed-off-by: John Strunk <jstrunk@redhat.com>